Change singleton behavior of HTTPServicesRegistry and WebSocketServicesRegistry
