+++ modules/ballerina-core/src/main/java/org/ballerinalang/connector/api/Service.java
@@ -40,6 +40,13 @@ public interface Service {
     String getPackage();
 
     /**
+     * This method returns the package of the protocol bound to the service.
+     *
+     * @return package of the protocol bound to the service.
+     */
+    String getProtocolPackage();
+
+    /**
      * This method will return annotation for the given package path and annotation name.
      *
      * @param pkgPath of the annotation.
