+++ modules/ballerina-http/src/main/java/org/ballerinalang/net/http/HTTPServicesRegistry.java
@@ -22,15 +22,23 @@ package org.ballerinalang.net.http;
 import org.ballerinalang.connector.api.AnnAttrValue;
 import org.ballerinalang.connector.api.Annotation;
 import org.ballerinalang.connector.api.BallerinaConnectorException;
+import org.ballerinalang.connector.api.ParamDetail;
+import org.ballerinalang.connector.api.Resource;
+import org.ballerinalang.net.uri.DispatcherUtil;
+import org.ballerinalang.net.uri.URITemplateException;
 import org.ballerinalang.net.ws.WebSocketServicesRegistry;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.wso2.transport.http.netty.config.ListenerConfiguration;
 
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.CopyOnWriteArrayList;
+import java.util.stream.Collectors;
 
 /**
  * This services registry holds all the services of HTTP + WebSocket.
@@ -44,13 +52,11 @@ public class HTTPServicesRegistry {
 
     // Outer Map key=interface, Inner Map key=basePath
     private final Map<String, Map<String, HttpService>> servicesInfoMap = new ConcurrentHashMap<>();
-    private static final HTTPServicesRegistry servicesRegistry = new HTTPServicesRegistry();
+    private CopyOnWriteArrayList<String> sortedServiceURIs = new CopyOnWriteArrayList<>();
+    private final WebSocketServicesRegistry webSocketServicesRegistry;
 
-    private HTTPServicesRegistry() {
-    }
-
-    public static HTTPServicesRegistry getInstance() {
-        return servicesRegistry;
+    public HTTPServicesRegistry(WebSocketServicesRegistry webSocketServicesRegistry) {
+        this.webSocketServicesRegistry = webSocketServicesRegistry;
     }
 
     /**
@@ -111,6 +117,7 @@ public class HTTPServicesRegistry {
             }
         }
         logger.info("Service deployed : " + service.getName() + " with context " + basePath);
+        postProcessService(service);
     }
 
     /**
@@ -138,6 +145,7 @@ public class HTTPServicesRegistry {
                 HttpConnectionManager.getInstance().closeIfLast(entryListenerInterface);
             }
         }
+        sortedServiceURIs.remove(service.getBasePath());
     }
 
     private String discoverBasePathFrom(HttpService service, Annotation annotation) {
@@ -175,6 +183,135 @@ public class HTTPServicesRegistry {
         String serviceName =
                 webSocketAnn.getAnnAttrValue(Constants.ANN_WEBSOCKET_ATTR_SERVICE_NAME).getStringValue().trim();
         String uri = basePath.concat(upgradePath);
-        WebSocketServicesRegistry.getInstance().registerServiceByName(serviceInterface, uri, serviceName);
+        webSocketServicesRegistry.registerServiceByName(serviceInterface, uri, serviceName);
+    }
+
+    private void postProcessService(HttpService httpService) {
+        CorsPopulator.populateServiceCors(httpService);
+        List<HttpResource> resources = new ArrayList<>();
+        for (Resource resource : httpService.getBalerinaService().getResources()) {
+            HttpResource httpResource = buildHttpResource(resource);
+            validateResourceSignature(httpResource);
+            try {
+                httpService.getUriTemplate().parse(httpResource.getPath(), httpResource,
+                                                   new HttpResourceElementFactory());
+            } catch (URITemplateException e) {
+                throw new BallerinaConnectorException(e.getMessage());
+            }
+            CorsPopulator.processResourceCors(httpResource, httpService);
+            resources.add(httpResource);
+        }
+        httpService.setResources(resources);
+        httpService.setAllAllowMethods(DispatcherUtil.getAllResourceMethods(httpService));
+        //basePath will get cached after registering service
+        sortedServiceURIs.add(httpService.getBasePath());
+        sortedServiceURIs.sort((basePath1, basePath2) -> basePath2.length() - basePath1.length());
+    }
+
+    private HttpResource buildHttpResource(Resource resource) {
+        HttpResource httpResource = new HttpResource(resource);
+        Annotation rConfigAnnotation = resource.getAnnotation(Constants.HTTP_PACKAGE_PATH,
+                                                              Constants.ANN_NAME_RESOURCE_CONFIG);
+        if (rConfigAnnotation == null) {
+            if (logger.isDebugEnabled()) {
+                logger.debug("resourceConfig not specified in the Resource, using default sub path");
+            }
+            httpResource.setPath(resource.getName());
+            return httpResource;
+        }
+        String subPath;
+        AnnAttrValue pathAttrVal = rConfigAnnotation.getAnnAttrValue(Constants.ANN_RESOURCE_ATTR_PATH);
+        if (pathAttrVal == null) {
+            if (logger.isDebugEnabled()) {
+                logger.debug("Path not specified in the Resource, using default sub path");
+            }
+            subPath = resource.getName();
+        } else {
+            subPath = pathAttrVal.getStringValue().trim();
+        }
+        if (subPath.isEmpty()) {
+            subPath = Constants.DEFAULT_BASE_PATH;
+        }
+        httpResource.setPath(subPath);
+
+        AnnAttrValue methodsAttrVal = rConfigAnnotation.getAnnAttrValue(Constants.ANN_RESOURCE_ATTR_METHODS);
+        if (methodsAttrVal != null) {
+            httpResource.setMethods(DispatcherUtil.getValueList(methodsAttrVal, null));
+        }
+        AnnAttrValue consumesAttrVal = rConfigAnnotation.getAnnAttrValue(Constants.ANN_RESOURCE_ATTR_CONSUMES);
+        if (consumesAttrVal != null) {
+            httpResource.setConsumes(DispatcherUtil.getValueList(consumesAttrVal, null));
+        }
+        AnnAttrValue producesAttrVal = rConfigAnnotation.getAnnAttrValue(Constants.ANN_RESOURCE_ATTR_PRODUCES);
+        if (producesAttrVal != null) {
+            httpResource.setProduces(DispatcherUtil.getValueList(producesAttrVal, null));
+        }
+        if (httpResource.getProduces() != null) {
+            List<String> subAttributeValues = httpResource.getProduces().stream()
+                    .map(mediaType -> mediaType.trim()
+                            .substring(0, mediaType.indexOf("/")))
+                    .distinct().collect(Collectors.toList());
+            httpResource.setProducesSubTypes(subAttributeValues);
+        }
+        return httpResource;
+    }
+
+    private void validateResourceSignature(HttpResource resource) {
+        List<ParamDetail> paramDetails = resource.getParamDetails();
+
+        if (paramDetails.size() < 2) {
+            throw new BallerinaConnectorException("resource signature parameter count should be more than two");
+        }
+
+        //Validate request parameter
+        ParamDetail reqParamDetail = paramDetails.get(0);
+        if (reqParamDetail == null) {
+            throw new BallerinaConnectorException("request parameter cannot be null");
+        }
+        if (reqParamDetail.getVarType().getPackagePath() == null
+                || !reqParamDetail.getVarType().getPackagePath().equals(Constants.PROTOCOL_PACKAGE_HTTP)
+                || !reqParamDetail.getVarType().getName().equals(Constants.REQUEST)) {
+            throw new BallerinaConnectorException("request parameter should be of type - "
+                                                          + Constants.PROTOCOL_PACKAGE_HTTP + ":" + Constants.REQUEST);
+        }
+
+        //validate response parameter
+        ParamDetail respParamDetail = paramDetails.get(1);
+        if (respParamDetail == null) {
+            throw new BallerinaConnectorException("response parameter cannot be null");
+        }
+        if (respParamDetail.getVarType().getPackagePath() == null
+                || !respParamDetail.getVarType().getPackagePath().equals(Constants.PROTOCOL_PACKAGE_HTTP)
+                || !respParamDetail.getVarType().getName().equals(Constants.RESPONSE)) {
+            throw new BallerinaConnectorException("response parameter should be of type - "
+                                                          + Constants.PROTOCOL_PACKAGE_HTTP + ":" + Constants.REQUEST);
+        }
+
+        //validate rest of the parameters
+        for (int i = 2; i < paramDetails.size(); i++) {
+            ParamDetail paramDetail = paramDetails.get(i);
+            if (!paramDetail.getVarType().getName().equals(Constants.TYPE_STRING)) {
+                throw new BallerinaConnectorException("incompatible resource signature parameter type");
+            }
+        }
+    }
+
+
+    public String findTheMostSpecificBasePath(String requestURIPath, Map<String, HttpService> services) {
+        for (Object key : sortedServiceURIs) {
+            if (!requestURIPath.toLowerCase().contains(key.toString().toLowerCase())) {
+                continue;
+            }
+            if (requestURIPath.length() <= key.toString().length()) {
+                return key.toString();
+            }
+            if (requestURIPath.charAt(key.toString().length()) == '/') {
+                return key.toString();
+            }
+        }
+        if (services.containsKey(Constants.DEFAULT_BASE_PATH)) {
+            return Constants.DEFAULT_BASE_PATH;
+        }
+        return null;
     }
 }
