+++ modules/ballerina-http/src/main/java/org/ballerinalang/net/http/HttpDispatcher.java
@@ -23,10 +23,12 @@ import org.ballerinalang.connector.api.ParamDetail;
 import org.ballerinalang.model.values.BString;
 import org.ballerinalang.model.values.BStruct;
 import org.ballerinalang.model.values.BValue;
+import org.ballerinalang.net.uri.URIUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.wso2.transport.http.netty.message.HTTPCarbonMessage;
 
+import java.net.URI;
 import java.util.List;
 import java.util.Map;
 
@@ -38,6 +40,56 @@ import java.util.Map;
 public class HttpDispatcher {
 
     private static final Logger breLog = LoggerFactory.getLogger(HttpDispatcher.class);
+
+    public static HttpService findService(HTTPServicesRegistry servicesRegistry, HTTPCarbonMessage cMsg) {
+
+        try {
+            String interfaceId = getInterface(cMsg);
+            Map<String, HttpService> servicesOnInterface = servicesRegistry.getServicesInfoByInterface(interfaceId);
+            if (servicesOnInterface == null) {
+                throw new BallerinaConnectorException("No services found for interface : " + interfaceId);
+            }
+            String uriStr = (String) cMsg.getProperty(org.wso2.carbon.messaging.Constants.TO);
+            //replace multiple slashes from single slash if exist in request path to enable
+            // dispatchers when request path contains multiple slashes
+            URI requestUri = URI.create(uriStr.replaceAll("//+", Constants.DEFAULT_BASE_PATH));
+            if (requestUri == null) {
+                throw new BallerinaConnectorException("uri not found in the message or found an invalid URI.");
+            }
+
+            // Most of the time we will find service from here
+            String basePath =
+                    servicesRegistry.findTheMostSpecificBasePath(requestUri.getPath(), servicesOnInterface);
+            HttpService service = servicesOnInterface.get(basePath);
+            if (service == null) {
+                cMsg.setProperty(Constants.HTTP_STATUS_CODE, 404);
+                throw new BallerinaConnectorException("no matching service found for path : " + uriStr);
+            }
+
+            String subPath = URIUtil.getSubPath(requestUri.getPath(), basePath);
+            cMsg.setProperty(Constants.BASE_PATH, basePath);
+            cMsg.setProperty(Constants.SUB_PATH, subPath);
+            cMsg.setProperty(Constants.QUERY_STR, requestUri.getQuery());
+            //store query params comes with request as it is
+            cMsg.setProperty(Constants.RAW_QUERY_STR, requestUri.getRawQuery());
+
+            return service;
+        } catch (Throwable e) {
+            throw new BallerinaConnectorException(e.getMessage());
+        }
+    }
+
+    protected static String getInterface(HTTPCarbonMessage cMsg) {
+        String interfaceId = (String) cMsg.getProperty(org.wso2.carbon.messaging.Constants.LISTENER_INTERFACE_ID);
+        if (interfaceId == null) {
+            if (breLog.isDebugEnabled()) {
+                breLog.debug("Interface id not found on the message, hence using the default interface");
+            }
+            interfaceId = Constants.DEFAULT_INTERFACE;
+        }
+
+        return interfaceId;
+    }
 
     public static void handleError(HTTPCarbonMessage cMsg, Throwable throwable) {
         String errorMsg = throwable.getMessage();
@@ -57,23 +109,17 @@ public class HttpDispatcher {
      * @param httpCarbonMessage incoming message.
      * @return matching resource.
      */
-    public static HttpResource findResource(HTTPCarbonMessage httpCarbonMessage) {
+    public static HttpResource findResource(HTTPServicesRegistry servicesRegistry,
+                                            HTTPCarbonMessage httpCarbonMessage) {
         HttpResource resource = null;
         String protocol = (String) httpCarbonMessage.getProperty(org.wso2.carbon.messaging.Constants.PROTOCOL);
         if (protocol == null) {
             throw new BallerinaConnectorException("protocol not defined in the incoming request");
         }
 
-        // Find the Service Dispatcher
-        HttpServerConnector serverConnector = (HttpServerConnector) ConnectorUtils
-                .getBallerinaServerConnector(Constants.PROTOCOL_PACKAGE_HTTP);
-        if (serverConnector == null) {
-            throw new BallerinaConnectorException("no service dispatcher available to handle protocol: " + protocol);
-        }
-
         try {
             // Find the Service TODO can be improved
-            HttpService service = serverConnector.findService(httpCarbonMessage);
+            HttpService service = HttpDispatcher.findService(servicesRegistry, httpCarbonMessage);
             if (service == null) {
                 throw new BallerinaConnectorException("no Service found to handle the service request");
                 // Finer details of the errors are thrown from the dispatcher itself, Ideally we shouldn't get here.
