+++ modules/ballerina-http/src/main/java/org/ballerinalang/net/http/HttpUtil.java
@@ -440,10 +440,11 @@ public class HttpUtil {
      *
      * @throws BallerinaConnectorException
      */
-    public static void startPendingHttpConnectors() throws BallerinaConnectorException {
+    public static void startPendingHttpConnectors(HttpServerConnector httpServerConnector)
+            throws BallerinaConnectorException {
         try {
             // Starting up HTTP Server connectors
-            HttpConnectionManager.getInstance().startPendingHTTPConnectors();
+            HttpConnectionManager.getInstance().startPendingHTTPConnectors(httpServerConnector);
         } catch (ServerConnectorException e) {
             throw new BallerinaConnectorException(e);
         }
