+++ modules/ballerina-test/src/test/java/org/ballerinalang/test/services/testutils/Services.java
@@ -20,11 +20,13 @@ package org.ballerinalang.test.services.testutils;
 
 
 import org.ballerinalang.connector.api.ConnectorFuture;
+import org.ballerinalang.connector.api.ConnectorUtils;
 import org.ballerinalang.connector.api.Executor;
 import org.ballerinalang.model.values.BValue;
 import org.ballerinalang.net.http.Constants;
 import org.ballerinalang.net.http.HttpDispatcher;
 import org.ballerinalang.net.http.HttpResource;
+import org.ballerinalang.net.http.HttpServerConnector;
 import org.wso2.transport.http.netty.message.HTTPCarbonMessage;
 
 import java.util.Collections;
@@ -38,9 +40,11 @@ import java.util.Map;
 public class Services {
 
     public static HTTPCarbonMessage invokeNew(HTTPTestRequest request) {
+        HttpServerConnector httpServerConnector =
+                (HttpServerConnector) ConnectorUtils.getBallerinaServerConnector(Constants.HTTP_PACKAGE_PATH);
         TestHttpFutureListener futureListener = new TestHttpFutureListener(request);
         request.setFutureListener(futureListener);
-        HttpResource resource = HttpDispatcher.findResource(request);
+        HttpResource resource = HttpDispatcher.findResource(httpServerConnector.getHttpServicesRegistry(), request);
         if (resource == null) {
             return futureListener.getResponseMsg();
         }
