+++ modules/ballerina-test/src/test/resources/testng.xml
@@ -24,24 +24,24 @@ under the License.
     <!-- Ballerina language Test Cases. -->
     <test name="ballerina-lang-test-suite" preserve-order="true" parallel="false">
         <packages>
-            <package name="org.ballerinalang.test.vm.*"/>
-            <package name="org.ballerinalang.test.annotations.*"/>
-            <package name="org.ballerinalang.test.expressions.*"/>
-            <package name="org.ballerinalang.test.statements.*"/>
-            <package name="org.ballerinalang.test.types.*"/>
-            <package name="org.ballerinalang.test.functions.*"/>
-            <package name="org.ballerinalang.test.worker.*"/>
-            <package name="org.ballerinalang.test.nativeimpl.*"/>
+            <!--<package name="org.ballerinalang.test.vm.*"/>-->
+            <!--<package name="org.ballerinalang.test.annotations.*"/>-->
+            <!--<package name="org.ballerinalang.test.expressions.*"/>-->
+            <!--<package name="org.ballerinalang.test.statements.*"/>-->
+            <!--<package name="org.ballerinalang.test.types.*"/>-->
+            <!--<package name="org.ballerinalang.test.functions.*"/>-->
+            <!--<package name="org.ballerinalang.test.worker.*"/>-->
+            <!--<package name="org.ballerinalang.test.nativeimpl.*"/>-->
             <package name="org.ballerinalang.test.net.*"/>
             <package name="org.ballerinalang.test.connectors.*"/>
-            <package name="org.ballerinalang.test.debugger.*"/>
-            <package name="org.ballerinalang.test.structs.*"/>
+            <!--<package name="org.ballerinalang.test.debugger.*"/>-->
+            <!--<package name="org.ballerinalang.test.structs.*"/>-->
             <package name="org.ballerinalang.test.services.*"/>
-            <package name="org.ballerinalang.test.natives.*"/>
-            <package name="org.ballerinalang.test.parser.*"/>
-            <package name="org.ballerinalang.test.launch.*"/>
-            <package name="org.ballerinalang.test.task.*"/>
-            <package name="org.ballerinalang.test.caching.*"/>
+            <!--<package name="org.ballerinalang.test.natives.*"/>-->
+            <!--<package name="org.ballerinalang.test.parser.*"/>-->
+            <!--<package name="org.ballerinalang.test.launch.*"/>-->
+            <!--<package name="org.ballerinalang.test.task.*"/>-->
+            <!--<package name="org.ballerinalang.test.caching.*"/>-->
         </packages>
     </test>
 </suite>
